+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 1.2 (unreleased)
 ----------------
 
+- Clean last database result from connections on pool checkin.
+
 - Close the database connections even if exceptions occurred.
 
 1.1 (2014-10-27)